修复构建失败
