+++ lib/cpp_native/bindings/danmaku_parser_io.dart
@@ -5,8 +5,8 @@ import 'package:flutter/foundation.dart';
 
 import '../native_arena.dart';
 import '../native_library.dart';
-import '../types/native_types.dart';
-import '../types/native_result.dart';
+import '../types/native_types_io.dart';
+import '../types/native_result_io.dart';
 import 'native_bindings.dart';
 
 /// C++ 弹幕解析器的 Dart FFI 封装
