+++ lib/cpp_native/bindings/similarity_engine_io.dart
@@ -4,8 +4,8 @@ import 'package:ffi/ffi.dart';
 import 'package:flutter/foundation.dart';
 
 import '../native_library.dart';
-import '../types/native_types.dart';
-import '../types/native_result.dart';
+import '../types/native_types_io.dart';
+import '../types/native_result_io.dart';
 import 'native_bindings.dart';
 
 /// 批量查重结果
