+++ lib/cpp_native/types/native_subtitle_types_web.dart
@@ -1,0 +1,11 @@
+/// Web stub for native subtitle types.
+///
+/// Provides the same public API as [native_subtitle_types_io.dart] (renamed
+/// from native_subtitle_types.dart) without any `dart:ffi` / `package:ffi`
+/// imports, so it compiles on Web. On Web these types are never actually used.
+
+/// Placeholder — on Web the native subtitle parser is never used.
+class NpSubtitleEntry {}
+
+/// Placeholder — on Web the native subtitle parser is never used.
+class NpSubtitleParseResult {}
